jinop

+++ backend/src/main/java/jinop/model/Client.java
@@ -13,7 +13,7 @@ public class Client implements Serializable {
     private String password;
     private BigDecimal money;
     private String name;
-    private Integer num;
+    private Long num;
     private String address;
     private Integer paypass;
     private Integer viplevel;
@@ -49,11 +49,11 @@ public class Client implements Serializable {
 
     public void setName(String name) { this.name = name; }
 
-    public Integer getNum() {
+    public Long getNum() {
         return num;
     }
 
-    public void setNum(Integer num) { this.num = num; }
+    public void setNum(Long num) { this.num = num; }
 
     public String getAddress() {
         return address;
